requirements.txt
# Core AI/ML
huggingface-hub>=0.20.0
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
upskill>=0.1.0

# LLM Inference
openai>=1.0.0

# Document Processing
pymupdf>=1.23.0
pdfplumber>=0.10.0
python-docx>=1.0.0

# Web Scraping & APIs
beautifulsoup4>=4.12.0
playwright>=1.40.0
requests>=2.31.0
aiohttp>=3.9.0

# Vector Store
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# UI
streamlit>=1.30.0
plotly>=5.18.0
matplotlib>=3.8.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
tqdm>=4.66.0
PyYAML>=6.0

# Development
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
